# Use faster strncmp implementation