--- a/src/init.c
+++ b/src/init.c
@@ -47,6 +47,41 @@ jl_array_t *jl_module_init_order;
 
 JL_DLLEXPORT size_t jl_page_size;
 
+int strncmp_fast( const char *ptr0, const char *ptr1, int len ){
+  int fast = len/sizeof(size_t) + 1;
+  int offset = (fast-1)*sizeof(size_t);
+  int current_block = 0;
+
+  if( len <= sizeof(size_t)){ fast = 0; }
+
+
+  size_t *lptr0 = (size_t*)ptr0;
+  size_t *lptr1 = (size_t*)ptr1;
+
+  while( current_block < fast ){
+    if( (lptr0[current_block] ^ lptr1[current_block] )){
+      int pos;
+      for(pos = current_block*sizeof(size_t); pos < len ; ++pos ){
+        if( (ptr0[pos] ^ ptr1[pos]) || (ptr0[pos] == 0) || (ptr1[pos] == 0) ){
+          return  (int)((unsigned char)ptr0[pos] - (unsigned char)ptr1[pos]);
+          }
+        }
+      }
+
+    ++current_block;
+    }
+
+  while( len > offset ){
+    if( (ptr0[offset] ^ ptr1[offset] )){
+      return (int)((unsigned char)ptr0[offset] - (unsigned char)ptr1[offset]);
+      }
+    ++offset;
+    }
+
+
+  return 0;
+}
+
 void jl_init_stack_limits(int ismaster, void **stack_lo, void **stack_hi)
 {
 #ifdef _OS_WINDOWS_
@@ -637,7 +672,7 @@ JL_DLLEXPORT int jl_is_file_tracked(jl_sym_t *path)
 {
     const char* path_ = jl_symbol_name(path);
     int tpath_len = strlen(jl_options.tracked_path);
-    return (strlen(path_) >= tpath_len) && (strncmp(path_, jl_options.tracked_path, tpath_len) == 0);
+    return (strlen(path_) >= tpath_len) && (strncmp_fast(path_, jl_options.tracked_path, tpath_len) == 0);
 }
 
 static void jl_set_io_wait(int v)