I came across this that claims to be a faster strncmp implementation and thought I'd see if there were any gains to be had.
https://mgronhol.github.io/fast-strcmp/

MacOS
master
```
julia> @btime run(`./julia --startup-file=no -e "1"`)
  205.421 ms (49 allocations: 1.73 KiB)
```
this PR
```
julia> @btime run(`./julia --startup-file=no -e "1"`)
  196.657 ms (49 allocations: 1.73 KiB)
``` 
The timing gain appears to be repeatable on my system.

I guess most of the gains here are from the `strncmp` in `symtab_lookup`.

Marking as draft as I haven't reviewed the implementation, and can't find a license.

___

Gains aren't apparent on ubuntu
master `75.763 ms (49 allocations: 1.77 KiB)`
this PR `76.883 ms (49 allocations: 1.73 KiB)`