@inkydragon 2022-06-26
https://github.com/JuliaLang/julia/blob/a60c76ea57de012b04ed7af6affe1d133c06cbda/src/debuginfo.cpp#L402-L406

This file already included `julia.h`, so there is no problem using the new implementation.
https://github.com/JuliaLang/julia/blob/a60c76ea57de012b04ed7af6affe1d133c06cbda/src/debuginfo.cpp#L25