@PallHaraldsson 2022-06-27
```
julia> @btime run(`./julia --startup-file=no -e "1"`)
  196.657 ms (49 allocations: 1.73 KiB)
```

For me running master for yesterday takes 158.863 ms

```
Gains aren't apparent on ubuntu
master 75.763 ms (49 allocations: 1.77 KiB)
this PR 76.883 ms (49 allocations: 1.73 KiB)
```

It's unclear what you are timing there, `run` ./julia also? Can't be since too fast (relative to Julia startup, and also unusually large speedup across distros if you were timing the same thing)? And is that ./julia the one compiled with your PR, or is the PR just used in the julia where you did the `@btime` from?