@yuyichao 2022-06-27
Do note that the article linked is 8 years old and one of the machine tested is at least 13 years old (according to wiki, 64bit pentium 4 was discontinued 13 years ago).

Given what glibc does, I highly doubt you can get better performance on new machines without processor feature dispatch.

This would be an even bigger issue on latest ARM machines that can efficiently vectorize functions like these with SVE.